requirements.txt
aiohappyeyeballs==2.3.4
aiohttp==3.10.1
aiosignal==1.3.1
aiosqlite==0.20.0
annotated-types==0.7.0
anthropic==0.32.0
anyio==4.4.0
async-timeout==4.0.3
attrs==24.1.0
certifi==2024.7.4
charset-normalizer==3.3.2
colorama==0.4.6
dataclasses-json==0.6.7
defusedxml==0.7.1
distro==1.9.0
exceptiongroup==1.2.2
faiss-cpu==1.8.0.post1
filelock==3.15.4
frozenlist==1.4.1
fsspec==2024.6.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.24.5
idna==3.7
jiter==0.5.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.2.12
langchain-anthropic==0.1.22
langchain-community==0.2.11
langchain-core==0.2.28
langchain-experimental==0.0.64
langchain-openai==0.1.20
langchain-text-splitters==0.2.2
langchainhub==0.1.20
langgraph==0.2.2
langgraph-checkpoint==1.0.2
langsmith==0.1.96
marshmallow==3.21.3
multidict==6.0.5
mypy-extensions==1.0.0
numpy==1.26.4
openai==1.39.0
orjson==3.10.6
packaging==24.1
pandas==2.2.2
pydantic==2.8.2
pydantic_core==2.20.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
regex==2024.7.24
requests==2.32.3
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.31
tenacity==8.5.0
tiktoken==0.7.0
tokenizers==0.19.1
tqdm==4.66.5
types-requests==2.32.0.20240712
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
yarl==1.9.4